refactor(server): update entities with relations and types
