+++ server/src/entities/Favorite.ts
@@ -1,19 +1,20 @@
-import { IsString } from 'class-validator';
+import { Field, ObjectType } from 'type-graphql';
+import { Column, Entity, ManyToOne } from 'typeorm';
+
+import BaseModels from 'entities/base/BaseModels';
 import Member from 'entities/Member';
 import Project from 'entities/Project';
-import { Field, ObjectType } from 'type-graphql';
-import { BaseEntity, Column, Entity, Index, ManyToOne } from 'typeorm';
 
 @Entity()
 @ObjectType()
-export default class Favorite extends BaseEntity {
+export default class Favorite extends BaseModels {
 	@Column()
 	@Field()
-	@ManyToOne(() => Project, { eager: true })
+	@ManyToOne(() => Project, (project) => project.id, { eager: true })
 	project: Project;
 
 	@Column()
 	@Field()
-	@ManyToOne(() => Member, { eager: true })
+	@ManyToOne(() => Member, (member) => member.id, { eager: true })
 	member: Member;
 }
