+++ server/src/entities/File.ts
@@ -1,12 +1,13 @@
 import { IsBoolean, IsDate, IsString } from 'class-validator';
 import { Field, ObjectType } from 'type-graphql';
-import { BaseEntity, Column, DeleteDateColumn, Entity, ManyToOne } from 'typeorm';
+import { Column, DeleteDateColumn, Entity, ManyToOne } from 'typeorm';
 
 import Project from 'entities/Project';
+import BaseModels from 'entities/base/BaseModels';
 
 @Entity()
 @ObjectType()
-export default class File extends BaseEntity {
+export default class File extends BaseModels {
 	@Column()
 	@Field()
 	@IsString()
@@ -19,7 +20,7 @@ export default class File extends BaseEntity {
 
 	@Column()
 	@Field()
-	@ManyToOne(() => Project, { eager: true })
+	@ManyToOne(() => Project, (project) => project.files, { eager: true })
 	project: Project;
 
 	@Column('boolean', { default: false })
